PA3/requirement.txt
Flask
flask_pymongo
bson
flask-graphql
pymongoexplain
pymysql